refactor: split workspace rest resource helpers
