+++ .sampo/changesets/issue-537-workspace-rest-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split workspace REST resource scaffolding into focused source-emitter and anchor-repair helpers while preserving the public add command surface.
