+++ packages/wp-typia-project-tools/tests/cli-add-workspace-boundaries.test.ts
@@ -17,6 +17,14 @@ test("cli-add-workspace delegates asset and rest-resource workflows to focused h
 		path.join(runtimeRoot, "cli-add-workspace-rest.ts"),
 		"utf8",
 	);
+	const restAnchorsSource = fs.readFileSync(
+		path.join(runtimeRoot, "cli-add-workspace-rest-anchors.ts"),
+		"utf8",
+	);
+	const restSourceEmittersSource = fs.readFileSync(
+		path.join(runtimeRoot, "cli-add-workspace-rest-source-emitters.ts"),
+		"utf8",
+	);
 
 	expect(addWorkspaceSource).toContain('from "./cli-add-workspace-assets.js"');
 	expect(addWorkspaceSource).toContain('from "./cli-add-workspace-rest.js"');
@@ -40,7 +48,15 @@ test("cli-add-workspace delegates asset and rest-resource workflows to focused h
 	expect(assetsSource).toContain("export async function runAddPatternCommand(");
 	expect(assetsSource).toContain("export async function runAddBindingSourceCommand(");
 	expect(assetsSource).toContain("export async function runAddEditorPluginCommand(");
-	expect(restSource).toContain("function buildRestResourceTypesSource(");
-	expect(restSource).toContain("async function ensureRestResourceBootstrapAnchors(");
+	expect(restSource).toContain('from "./cli-add-workspace-rest-anchors.js"');
+	expect(restSource).toContain('from "./cli-add-workspace-rest-source-emitters.js"');
+	expect(restSource).not.toContain("function buildRestResourceTypesSource(");
+	expect(restSource).not.toContain("async function ensureRestResourceBootstrapAnchors(");
+	expect(restSource).toContain("function buildRestResourcePhpSource(");
 	expect(restSource).toContain("export async function runAddRestResourceCommand(");
+	expect(restSourceEmittersSource).toContain("function buildRestResourceTypesSource(");
+	expect(restSourceEmittersSource).toContain("function buildRestResourceApiSource(");
+	expect(restSourceEmittersSource).toContain("function buildRestResourceDataSource(");
+	expect(restAnchorsSource).toContain("async function ensureRestResourceBootstrapAnchors(");
+	expect(restAnchorsSource).toContain("async function ensureRestResourceSyncScriptAnchors(");
 });
